PICARD-1695: Allow inline editing of script names
Removed the separate input for the script name in the script editor and enabled inline editing in the script list. Added a context menu for the editing.
