+++ picard/ui/ui_options_script.py
@@ -29,7 +29,7 @@ class Ui_ScriptingOptionsPage(object):
         self.splitter.setOrientation(QtCore.Qt.Horizontal)
         self.splitter.setChildrenCollapsible(False)
         self.splitter.setObjectName("splitter")
-        self.script_list = QtWidgets.QListWidget(self.splitter)
+        self.script_list = ScriptListWidget(self.splitter)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.MinimumExpanding, QtWidgets.QSizePolicy.Expanding)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(0)
@@ -42,14 +42,6 @@ class Ui_ScriptingOptionsPage(object):
         self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.formWidget)
         self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
         self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.script_name = QtWidgets.QLineEdit(self.formWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.script_name.sizePolicy().hasHeightForWidth())
-        self.script_name.setSizePolicy(sizePolicy)
-        self.script_name.setObjectName("script_name")
-        self.verticalLayout_2.addWidget(self.script_name)
         self.tagger_script = QtWidgets.QTextEdit(self.formWidget)
         self.tagger_script.setAcceptRichText(False)
         self.tagger_script.setObjectName("tagger_script")
@@ -60,9 +52,6 @@ class Ui_ScriptingOptionsPage(object):
         self.add_button = QtWidgets.QToolButton(self.enable_tagger_scripts)
         self.add_button.setObjectName("add_button")
         self.horizontalLayout.addWidget(self.add_button)
-        self.rename_button = QtWidgets.QToolButton(self.enable_tagger_scripts)
-        self.rename_button.setObjectName("rename_button")
-        self.horizontalLayout.addWidget(self.rename_button)
         spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
         self.horizontalLayout.addItem(spacerItem)
         self.move_up_button = QtWidgets.QToolButton(self.enable_tagger_scripts)
@@ -95,23 +84,19 @@ class Ui_ScriptingOptionsPage(object):
         self.retranslateUi(ScriptingOptionsPage)
         self.add_button.clicked.connect(ScriptingOptionsPage.add_script)
         self.tagger_script.textChanged.connect(ScriptingOptionsPage.live_update_and_check)
-        self.script_name.textChanged['QString'].connect(ScriptingOptionsPage.script_name_changed)
         self.script_list.itemSelectionChanged.connect(ScriptingOptionsPage.script_selected)
         self.remove_button.clicked.connect(ScriptingOptionsPage.delete_selected_script)
-        self.rename_button.clicked.connect(ScriptingOptionsPage.rename_selected_script)
         QtCore.QMetaObject.connectSlotsByName(ScriptingOptionsPage)
         ScriptingOptionsPage.setTabOrder(self.enable_tagger_scripts, self.add_button)
         ScriptingOptionsPage.setTabOrder(self.add_button, self.script_list)
-        ScriptingOptionsPage.setTabOrder(self.script_list, self.script_name)
-        ScriptingOptionsPage.setTabOrder(self.script_name, self.tagger_script)
+        ScriptingOptionsPage.setTabOrder(self.script_list, self.tagger_script)
 
     def retranslateUi(self, ScriptingOptionsPage):
         _translate = QtCore.QCoreApplication.translate
         self.enable_tagger_scripts.setTitle(_("Tagger Script(s)"))
-        self.script_name.setPlaceholderText(_("Display Name"))
         self.add_button.setToolTip(_("Add new script"))
         self.add_button.setText(_("Add new script"))
-        self.rename_button.setText(_("Rename script"))
         self.move_up_button.setToolTip(_("Move script up"))
         self.move_down_button.setToolTip(_("Move script down"))
         self.remove_button.setText(_("Remove script"))
+from picard.ui.widgets.scriptlistwidget import ScriptListWidget
